Add a SensorsMixin to remove duplicate code
